\name{MSstatsBioData-package}
\alias{MSstatsBioData-package}
\alias{MSstatsBioData}
\docType{package}

\title{Datasets of published biological studies with DDA or SRM experiments}

\description{
Provides the peak intensity data for detecting differentially abundant proteins between groups. Seven datasets from published biological investigations are available. (see \code{\link{DDA_cardio}}, \code{\link{SRM_yeast}}, \code{\link{SRM_ovarian}}, \code{\link{SRM_crc_training}}, \code{\link{SRM_crc_validation}}, \code{\link{SRM_mpm_training}}, \code{\link{SRM_mpm_validation}})
}

\details{
All datasets was processed as described in original reference. They were reformatted as MSstats required format.

To view the example workflows, type \code{browseVignettes("MSstatsBioData")}.
}

\author{
Meena Choi

Maintainer: Meena Choi <mnchoi67@gmail.com>
}

\references{
Clough, T. et al. (2009) Protein quantification in label-free LC-MS experiments. \emph{J. Proteome Res.}, 8, 5275–5284.

Picotti, P. et ak. (2009) Full dynamic range proteome analysis of S. cerevisiae by targeted proteomics. \emph{Cell}, 138, 795–806.

Huttenhain, R. et al. (2012) Reproducible quantification of cancer-associated proteins in body fluids using targeted proteomics. \emph{Sci. Tansl. Med.}, 4, 142ra94.

Cerciello, F. et al. (2013) Identification of a seven glycopeptide signature for malignant pleural mesothelioma in human serum by selected reaction monitoring. \emph{Clin. Proteomics}, 10, 16.

Surinova, S. et al. (2015) Prediction of colorectal cancer diagnosis based on circulating plasma proteins. \emph{EMBO Mol. Med.}, 7, 1166–1178.
}

\keyword{package, MSstatsBioData}
