\name{SRM_crc_validation}
\docType{data}

\alias{data:SRM_crc_validation}
\alias{SRM_crc_validation}


\title{The validation set from a study for subjects with colorectal cancer}
\description{
72 proteins, including two standard proteins, AIAG-Bovine and FETUA-Bovine, were targeted for plasma samples with SRM with isotope labeled reference peptides in order to identify candidate protein biomarker for non-invasive detection of CRC. The validation cohort had 67 subjects in controls, and 202 subject with different clinical stages of CRC. Each sample for subject was measured in a single injection without technical replicate. The validation cohort was processed with MultiQuant 1.2. NAs should be considered as censored missing. Two standard proteins can be removed for statistical analysis.
}
\usage{
data(SRM_crc_validation)
}
\format{

    \code{SRM_crc_validation} is a data.frame.

}
\references{
Surinova, S. et al. (2015) Prediction of colorectal cancer diagnosis based on circulating plasma proteins. \emph{EMBO Mol. Med.}, 7, 1166–1178.
}
\examples{
data(SRM_crc_validation)
}
\keyword{datasets}
